Rework handling interfaces and VLANs
Forcing IF_WAN and IF_ROUTER to be physical interfaces means
more exotic devices like bridges can't be used. It was also
realized that VLANs are not necessary if on VLAN ID 0 but
may still be necessary if on other VLAN IDs.

Remove the physical device check and allow configuring any
VLAN, or none, as IF_VLAN. Do so in a manner that keeps
compatibility and user transparency with respect to old
config file/option syntax.