Fix setting MAC addresses
IF_WAN and IF_VLAN can now be e.g. a bridge and a VLAN
subinterface on the bridge. Or, both can point to the same
physical or virtual interface.

In any case, the old code was simply assuming the existence
of IF_WAN.0 and setting its MAC, leaving the MAC of IF_WAN
untouched. IF_WAN itself should've been unused so this
probably would've been fine then, but now we certainly do
want to set both.

(Also, when the utility used to set MAC was changed from
ifconfig to ip, the cmdline syntax wasn't, so the feature
was entirely nonfunctional on this fork. Oops.)

Actually set MAC of IF_WAN (and also of IF_VLAN if used).
Do it with whichever of ip or ifconfig is present.